refactor(cli): moved get account args to common mod
